+++ src/Category_Colors/CSS/Extras.php
@@ -93,6 +93,33 @@ class Extras {
 	}
 
 	/**
+	 * Add new featured event styling for TEC/ECP 6.x
+	 *
+	 * @param array $options Array of options.
+	 *
+	 * @return void
+	 */
+	public static function add_new_featured_event( $options ) {
+		$css   = [];
+		$css[] = '.tribe-common--breakpoint-medium.tribe-events .tribe-events-calendar-list__event-datetime-featured-text,';
+		$css[] = '.tribe-events-pro .tribe-events-pro-photo__event-datetime-featured-text,';
+		$css[] = '.tribe-events-calendar-day__event-datetime-featured-text,';
+		$css[] = '.tribe-common .tribe-common-c-svgicon';
+		$css[] = "{ color: {$options['featured-event']} !important; }";
+
+		$css[] = '';
+
+		$css[] = '.tribe-events-calendar-list__event-date-tag-datetime:after, ';
+		$css[] = '.tribe-events-widget-events-list__event-date-tag-datetime:after, ';
+		$css[] = '.tribe-events-pro .tribe-events-pro-week-grid__event--featured .tribe-events-pro-week-grid__event-link-inner:before,';
+		$css[] = '.tribe-events .tribe-events-calendar-day__event--featured:after,';
+		$css[] = '.tribe-events .tribe-events-calendar-month__calendar-event--featured:before';
+		$css[] = "{ background-color: {$options['featured-event']} !important; }";
+
+		self::echo_css( $css );
+	}
+
+	/**
 	 * Add mobile CSS
 	 *
 	 * @return void