+++ src/views/category.css.php
@@ -39,6 +39,8 @@ use Fragen\Category_Colors\CSS\V2_Views;
 
 <?php Pro::fix_pro_spacer_background(); ?>
 
+<?php Extras::add_new_featured_event( $options ); ?>
+
 <?php
 if ( empty( $teccc->terms ) && ! empty( $options['terms'] ) ) {
 	$teccc->terms = $options['terms'];