+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 7.1.0 / 2022-10-12
 * rename title to correspond to The Events Calendar standard
 * retructure CSS additions to not end with comma, added in `category.css.php`
 * fix empty spacer classes to be transparent